requirements.txt
Flask==2.1.0
Pillow==9.5.0
pillow-heif==0.10.0
Werkzeug==2.2.3